Move Currency to Libplanet.Assets namespace
