+++ Libplanet.Tests/Assets/CurrencyTest.cs
@@ -3,11 +3,12 @@ using System.Collections.Immutable;
 using System.IO;
 using System.Runtime.Serialization.Formatters.Binary;
 using System.Security.Cryptography;
+using Libplanet.Assets;
 using Libplanet.Crypto;
 using Xunit;
 using static Libplanet.Tests.TestUtils;
 
-namespace Libplanet.Tests
+namespace Libplanet.Tests.Assets
 {
     public class CurrencyTest
     {
